Fix GUI helper imports for CI without PySide6
- move mask/commit helpers to api_key_helpers.py without Qt dependency

- normalize duplicate key comparison across whitespace variants

- add regression tests for failed check and already-applied gates
