+++ gui_qt/api_key_dialog.py
@@ -15,27 +15,7 @@ from PySide6.QtWidgets import (
     QWidget,
 )
 
-
-def mask_api_key(key: str) -> str:
-    stripped = key.strip()
-    if not stripped:
-        return "(空)"
-    suffix = stripped[-4:] if len(stripped) > 4 else "****"
-    return f"********{suffix}"
-
-
-def commit_pending_key(keys: list[str], pending_key: str) -> tuple[list[str], str | None]:
-    """Merge a non-empty input value into the key list before saving.
-
-    Returns ``(updated_keys, error_message)``. ``error_message`` is set when the
-    pending value duplicates an existing key.
-    """
-    pending = pending_key.strip()
-    if not pending:
-        return list(keys), None
-    if pending in keys:
-        return list(keys), "duplicate"
-    return [*keys, pending], None
+from .api_key_helpers import commit_pending_key, mask_api_key
 
 
 class ApiKeyDialog(QDialog):
